requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-parler==2.2
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
permission==0.4.1
Pillow==8.1.2
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
